Validations

+++ src/main/java/com/pranitpatil/entity/Stock.java
@@ -1,6 +1,7 @@
 package com.pranitpatil.entity;
 
 import javax.persistence.*;
+import java.math.BigDecimal;
 
 @Entity
 @Table
@@ -14,7 +15,7 @@ public class Stock extends AuditableEntity {
     private String name;
 
     @Column
-    private double price;
+    private BigDecimal price;
 
     public long getId() {
         return id;
@@ -32,11 +33,10 @@ public class Stock extends AuditableEntity {
         this.name = name;
     }
 
-    public double getPrice() {
+    public BigDecimal getPrice() {
         return price;
     }
-
-    public void setPrice(double price) {
+    public void setPrice(BigDecimal price) {
         this.price = price;
     }
 }
